Compilation: fix clang warning, I hope, by making cast explicit.
It appears that the implicit cast that turns 65536 into 0 which
then gets stored in a b-byte field is correct.  This patch makes
the cast explicit.

+++ plugins/excel/ms-excel-write.c
@@ -4809,7 +4809,7 @@ excel_write_PAGE_BREAK (BiffPut *bp, GnmPageBreaks const *breaks)
 	guint8 *data;
 	GnmPageBreaks *manual_pbreaks = gnm_page_breaks_dup_non_auto_breaks (breaks);
 	GArray *details =  manual_pbreaks->details;
-	guint16 const maxima = manual_pbreaks->is_vert ? XLS_MaxRow_V8 : XLS_MaxCol;
+	guint16 const maxima = (guint16)(manual_pbreaks->is_vert ? XLS_MaxRow_V8 : XLS_MaxCol);
 
 	/* limit size to ensure no CONTINUE (do we need this ? ) */
 	if (((n = details->len)*step + 2 + 2) >= ms_biff_max_record_len (bp))
